+++ packages/logd/lib/src/handler/engine/async_handler.dart
@@ -1,0 +1,4 @@
+library;
+
+export 'async_handler_stub.dart'
+    if (dart.library.io) 'async_handler_native.dart';